backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-google-genai
langchain
langchain-community
langchain-core
gitpython
python-dotenv
pydantic
pytest
requests
